+++ TriviaServer/TriviaServer/DAO/Repositories/GameRepository.cs
@@ -19,8 +19,15 @@ namespace TriviaServer.DAO.Repositories
 
         public void Create(Game g)
         {
-            _context.Games.Add(g);
-            _context.SaveChanges();
+            if (_context.Games.Where(a => a.UniqueKey == g.UniqueKey).SingleOrDefault() != null)
+            {
+                _context.Games.Add(g);
+                _context.SaveChanges();
+            }
+            else
+            {
+                throw new Exception("Game already exists!");
+            }
         }
 
         public void Edit(Game g)
@@ -32,14 +39,30 @@ namespace TriviaServer.DAO.Repositories
         public void Delete(int id)
         {
             var game = _context.Games.SingleOrDefault(x => x.GameId == id);
-            _context.Games.Remove(game);
-            _context.SaveChanges();
+            if(game != null)
+            {
+                _context.Games.Remove(game);
+                _context.SaveChanges();
+            }
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+       
         }
 
         public IEnumerable<Game> GetGames()
         {
             var games = _context.Games.ToList();
-            return games;
+            if(games != null)
+            {
+                return games;
+            }
+            else
+            {
+                throw new Exception("No game was found!");
+            }
+           
         }
 
         public Game GetByID(int? id)
@@ -50,50 +73,97 @@ namespace TriviaServer.DAO.Repositories
             }
 
             var game = _context.Games.SingleOrDefault(p => p.GameId == id);
-
-            return game;
+            if (game != null)
+            {
+                return game;
+            }
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+               
         }
 
         public List<PlayerScore> GetPlayerAndScoreByGameRoomId(int gameRoomId)
         {
-            var players = new List<PlayerScore>();
-            _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
-                .ForEach(a => players.Add(new PlayerScore() { Name = a.PlayerName, Score = a.PlayerScore, GameroomId = a.GameroomId }));
+            if (_context.Games.Where(a => a.GameId == gameRoomId).SingleOrDefault() != null)
+            {
+                var players = new List<PlayerScore>();
+                _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
+                    .ForEach(a => players.Add(new PlayerScore() { Name = a.PlayerName, Score = a.PlayerScore, GameroomId = a.GameroomId }));
 
-            return players;
+                return players;
+            }
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+            
         }
 
         public List<PlayerName> GetPlayersByRoomId(int gameRoomId)
         {
-            var players = new List<PlayerName>();
-            _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
-                .ForEach(a => players.Add(new PlayerName() { Name = a.PlayerName}));
-            return players;
+            if (_context.Games.Where(a => a.GameId == gameRoomId).SingleOrDefault() != null)
+            {
+                var players = new List<PlayerName>();
+                _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
+                    .ForEach(a => players.Add(new PlayerName() { Name = a.PlayerName }));
+                return players;
+            }
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+
         }
 
         public double GetAverageScore(int gameRoomId)
         {
-            double sum = 0;
-            _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
-                .ForEach(a => sum += a.PlayerScore);
-            var numberOfPlayers = GetPlayersByRoomId(gameRoomId).Count;
-            return sum / numberOfPlayers;
+            var games = GetGames();
+            if (_context.Games.Where(a => a.GameId == gameRoomId).SingleOrDefault() != null)
+            {
+                double sum = 0;
+                _context.Players.Where(a => a.GameroomId == gameRoomId).ToList()
+                    .ForEach(a => sum += a.PlayerScore);
+                var numberOfPlayers = GetPlayersByRoomId(gameRoomId).Count;
+                return sum / numberOfPlayers;
+            }
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+           
         }
 
         public List<QuestionAnswers> GetQuestionsAndAnswersByGameRoomId(int gameRoomId)
         {
-            var questions = new List<QuestionAnswers>();
-            var categories = new List<Category>();
-            _context.Categories.Where(a => a.GameroomId == gameRoomId).ToList()
-                .ForEach(a => categories.Add(a));
-            foreach(Category c in categories)
+            var games = GetGames();
+            if (_context.Games.Where(a => a.GameId == gameRoomId).SingleOrDefault() != null)
             {
-                _context.Questions.Where(a => a.CategoryId == c.CategoryId).ToList()
-                    .ForEach(a => questions.Add(new QuestionAnswers() { QuestionText=a.QuestionText ,
-                        CorrectAnswer = a.CorrectAnswer, WrongAnswer1=a.WrongAnswer1, WrongAnswer2=a.WrongAnswer2
-                    ,WrongAnswer3=a.WrongAnswer3}));
+                var questions = new List<QuestionAnswers>();
+                var categories = new List<Category>();
+                _context.Categories.Where(a => a.GameroomId == gameRoomId).ToList()
+                    .ForEach(a => categories.Add(a));
+                foreach (Category c in categories)
+                {
+                    _context.Questions.Where(a => a.CategoryId == c.CategoryId).ToList()
+                        .ForEach(a => questions.Add(new QuestionAnswers()
+                        {
+                            QuestionText = a.QuestionText,
+                            CorrectAnswer = a.CorrectAnswer,
+                            WrongAnswer1 = a.WrongAnswer1,
+                            WrongAnswer2 = a.WrongAnswer2
+                        ,
+                            WrongAnswer3 = a.WrongAnswer3
+                        }));
+                }
+                return questions;
             }
-            return questions;
+            else
+            {
+                throw new Exception("Gameroom not found!");
+            }
+            
         }
